refatoring -> moved EnemiesLogic to controllers
